+++ lib/solid_objects.rb
@@ -26,11 +26,13 @@ require "solid_objects/actor_definition"
 require "solid_objects/application_write_guard"
 require "solid_objects/actor"
 require "solid_objects/reference"
+require "solid_objects/outcome"
 require "solid_objects/message_reference"
 require "solid_objects/administration_audit"
 require "solid_objects/redrive_task"
 require "solid_objects/redrive_manager"
 require "solid_objects/redrive_runner"
+require "solid_objects/dead_row"
 require "solid_objects/dead_letter_scope"
 require "solid_objects/dead_letter_manager"
 require "solid_objects/message_pruner"
@@ -228,7 +230,7 @@ module SolidObjects
       WakeUpAdapters.build(configuration.wake_up_adapter)
     rescue ArgumentError
       raise
-    rescue => error
+    rescue ActiveRecord::ActiveRecordError, SystemCallError, IOError => error
       unreachable_wake_up(error)
     end