+++ test/dummy/web_mount_check.rb
@@ -11,16 +11,10 @@ ENV["RAILS_ENV"] = "test"
 require_relative "config/environment"
 require "solid_objects/web"
 require "rack/mock_request"
-require_relative "../../db/migrate/20260805000000_create_solid_objects_tables"
-require_relative "../../db/migrate/20260806000000_add_state_revision_to_solid_objects_instances"
-require_relative "../../db/migrate/20260813000000_rename_message_dispatch_columns"
-require_relative "../../db/migrate/20260915000000_add_solid_objects_effect_recoveries"
+require "solid_objects/schema_bootstrap"
 
 ActiveRecord::Migration.verbose = false
-CreateSolidObjectsTables.new.migrate(:up)
-AddStateRevisionToSolidObjectsInstances.new.migrate(:up)
-RenameMessageDispatchColumns.new.migrate(:up)
-AddSolidObjectsEffectRecoveries.new.migrate(:up)
+SolidObjects::SchemaBootstrap.install
 
 instance = SolidObjects::Instance.create!(
   actor_type: "MountCheckActor",